feature: [be] 홈 화면 정보 조회

+++ src/main/java/com/yoganavi/live_lecture/repository/LectureScheduleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LectureScheduleRepository extends JpaRepository<LectureSchedule, Integer> {
+public interface LectureScheduleRepository extends JpaRepository<LectureSchedule, Long> {
 
 }
